requirements.txt
torch
torch_geometric
torch_scatter
torch_sparse
torch_cluster
torch_spline_conv
scikit-learn
pandas
numpy
matplotlib
networkx
rdkit-pypi
streamlit
